create new payment: fill info finished

+++ app/src/main/java/cz/brno/holan/jiri/hunggarkuenfinancials/backend/entities/products/OneTimeOnly.java
@@ -22,7 +22,7 @@ public class OneTimeOnly extends Product {
     public OneTimeOnly() {
     }
 
-    public OneTimeOnly(long id, String name, int value, int stock) {
+    public OneTimeOnly(long id, String name, float value, int stock) {
         super(id, name, value);
         detail = stock;
     }
